refactor: rename modal
