+++ src/components/Modals/AddToContestModal/index.ts
@@ -1,0 +1,3 @@
+import AddToContestModal from './AddToContestModal.container'
+
+export default AddToContestModal
