--- src/components/Modals/SubmitProjectModal/index.ts
@@ -1,3 +1,0 @@
-import SubmitProjectModal from './SubmitProjectModal.container'
-
-export default SubmitProjectModal
