+++ src/components/Modals/index.ts
@@ -1,4 +1,4 @@
 export { default as ShortcutsModal } from './ShortcutsModal'
 export { default as ContestModal } from './ContestModal'
 export { default as EditProjectModal } from './EditProjectModal'
-export { default as SubmitProjectModal } from './SubmitProjectModal'
+export { default as AddToContestModal } from './AddToContestModal'
